% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honest_weak_factors.R
\name{honest_weak_factors}
\alias{honest_weak_factors}
\title{Robust Estimation and Inference in Panels with Interactive Fixed Effects}
\usage{
honest_weak_factors(Y, X, R, Gamma_LS = NULL, alpha = 0.05)
}
\arguments{
\item{Y}{\eqn{N \times T} matrix of outcomes}

\item{X}{\eqn{K \times N \times T} tensor of regressors}

\item{R}{A positive integer, indicates the number of interactive fixed effects in the estimation}

\item{Gamma_LS}{(Optional) A preliminary LS estimate of the matrix of fixed effects; it will be computed if not provided}

\item{alpha}{(Optional) It determines the \code{1 - alpha} coverage of the constructed CI, where default is set to \code{alpha = 0.05}}
}
\value{
A list of results, where
\itemize{
\item \code{beta} is the point estimate
\item \code{bias} is the worst-case bias
\item \code{LB} is the lower bound of the \code{1 - alpha} CI
\item \code{UB} is the upper bound of the \code{1 - alpha} CI
\item \code{LB_unadj} is the lower bound of the \code{1 - alpha} CI without bias correction
\item \code{UB_unadj} is the upper bound of the \code{1 - alpha} CI without bias correction
\item \code{A} is the matrix of weights
}
}
\description{
This method considers estimation and inference for a regression coefficient in panels with interactive fixed effects (i.e., with a factor structure). As the previously developed estimators and confidence intervals (CIs) might be heavily biased and size-distorted when some of the factors are weak, this method has estimators with improved rates of convergence and bias-aware CIs that are uniformly valid regardless of whether the factors are strong or not.
}
