% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.honest_weak_factors.R
\name{summary.honest_weak_factors}
\alias{summary.honest_weak_factors}
\alias{print.summary.honest_weak_factors}
\title{Summarizing Robust Estimation and Inference in Panels with Interactive Fixed Effects}
\usage{
\method{summary}{honest_weak_factors}(object, ...)

\method{print}{summary.honest_weak_factors}(x, digits = 4, labels = TRUE, ...)
}
\arguments{
\item{object}{The \code{honest_weak_factors} fitted result of class \code{honest_weak_factors}.}

\item{...}{Confidence level \code{alpha} can be set for computing confidence intervals. Default is set to \eqn{0.05}.}

\item{x}{The summarized result of class \code{summary.ls_factor}.}

\item{digits}{Number of digits to use when printing. Default is set to \eqn{4}.}

\item{labels}{Option of whether or not to print labels in the summary. Default is set to \code{TRUE}. Only parameters estimates will be printed when it's set to \code{FALSE}.}
}
\value{
A list of results, where
\itemize{
\item \code{beta} is the debiased parameter estimate.
\item \code{se} is the standard errors of the parameters estimator.
\item \code{bias} is the worst-case bias of the estimator.
\item \code{CI} is the bias-aware confidence interval.
\item \code{CI_unadj} is the confidence interval without bias adjustment.
\item \code{A} is the choice of weight matrix.
}
}
\description{
\code{summary} method for object of class \code{honest_weak_factors} and returned object is of class {summary.honest_weak_factors}
}
\details{
Summarizing the fitted \code{honest_weak_factors} object and computing
\itemize{
\item Debiased parameter estimate.
\item Standard error.
\item Worst-case bias of the estimator.
\item With and without bias-aware confidence intervals.
}
}
