% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ls_factor.R
\name{summary.ls_factor}
\alias{summary.ls_factor}
\alias{print.summary.ls_factor}
\title{Summarizing Least Squares Estimation of Linear Panel Data Models with Interactive Fixed Effects}
\usage{
\method{summary}{ls_factor}(object, ...)

\method{print}{summary.ls_factor}(x, digits = 4, labels = TRUE, ...)
}
\arguments{
\item{object}{The \code{ls_factor} fitted result of class \code{ls_factor}.}

\item{...}{Confidence level \code{alpha} can be set for computing confidence intervals. Default is set to \eqn{0.05}.}

\item{x}{The summarized result of class \code{summary.ls_factor}.}

\item{digits}{Number of digits to use when printing. Default is set to \eqn{4}.}

\item{labels}{Option of whether or not to print labels in the summary. Default is set to \code{TRUE}. Only parameters estimates will be printed when it's set to \code{FALSE}.}
}
\value{
A list of results, where
\itemize{
\item \code{beta} is the table of parameters estimates under different bias correction schemes.
\item \code{CI} is the table of confidence intervals and their lengths under different settings.
\item \code{var_cov} are the variance-covariance matrices under different assumptions.
}
}
\description{
\code{summary} method for object of class \code{ls_factor} and returned object is of class {summary.ls_factor}
}
\details{
Summarizing the fitted \code{ls_factor} object and computing
\itemize{
\item Estimates under different bias correction schemes.
\item Variance-covariance matrix under different assumptions.
\item Confidence intervals and their lengths under different settings.
}
}
