% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honest_weak_factors.R
\name{honest_weak_factors}
\alias{honest_weak_factors}
\title{Robust Estimation and Inference in Panels with Interactive Fixed Effects}
\usage{
honest_weak_factors(
  Y,
  X,
  R,
  Gamma_LS = NULL,
  alpha = 0.05,
  clustered_se = FALSE,
  lambda_known = NA,
  f_known = NA,
  itermax = 75,
  reltol = 10^(-6)
)
}
\arguments{
\item{Y}{\eqn{N \times T} matrix of outcomes}

\item{X}{\eqn{K \times N \times T} tensor of regressors}

\item{R}{A positive integer, indicates the number of interactive fixed effects in the estimation. Note that this number does not include the known factors and loadings defined below}

\item{Gamma_LS}{(Optional) A preliminary LS estimate of the matrix of fixed effects; it will be computed if not provided}

\item{alpha}{(Optional) It determines the \code{1 - alpha} coverage of the constructed confidence interval, where default is set to \code{alpha = 0.05}}

\item{clustered_se}{(Optional) Whether or not performing clustered standard error, where default is set to \code{clustered_se = FALSE}}

\item{lambda_known}{(Optional) \eqn{N \times Rex1} matrix of known factor loadings, e.g., \code{lambda_known = matrix(rep(1, N), nrow = N, ncol = 1)} to control standard time dummies.
Default is set to \code{lambda_known = NA} (or equivalently \code{matrix(NA, nrow = N, ncol = 0)}), i.e., there is no known factor loadings}

\item{f_known}{(Optional) \eqn{T \times Rex2} matrix of known factor, e.g., \code{f_known = matrix(rep(1, T), nrow = T, ncol = 1)} to control standard individual specific fixed effects.
Default is set to \code{f_known = NA} (or equivalently \code{matrix(NA, nrow = T, ncol = 0)}), i.e., there is no known factor}

\item{itermax}{(Optional) Maximum iteration allowed while optimizing for the weights A, where default is set to \code{itermax = 50} for faster computation}

\item{reltol}{(Optional) Relative convergence tolerance for the optimization algorithm to stop if it is unable to reduce the value by \code{reltol * (abs(val) + reltol)} after \code{0.75 * itermax} steps.
, where default is set to \code{reltol = 10^(-4)} for faster computation}
}
\value{
A list of results, where
\itemize{
\item \code{beta} is the point estimate
\item \code{bias} is the worst-case bias
\item \code{LB} are the lower bounds of the \code{1 - alpha} confidence intervals, from number of weak factors being 0 to R
\item \code{UB} are the upper bounds of the \code{1 - alpha} confidence intervals, from number of weak factors being 0 to R
\item \code{A} is the matrix of weights
\item \code{parameter} is the list of input parameters, including \code{Gamma_LS}, \code{alpha}, and \code{clustered_se}
}
}
\description{
This method considers estimation and inference for a regression coefficient in panels with interactive fixed effects (i.e., with a factor structure). As the previously developed estimators and confidence intervals (CIs) might be heavily biased and size-distorted when some of the factors are weak, this method has estimators with improved rates of convergence and bias-aware CIs that are uniformly valid regardless of whether the factors are strong or not.
}
\details{
\strong{Disclaimer:} This function is the implementation of Armstrong, Weidner, and Zeleneev (2024, "Robust Estimation and Inference in Panels with Interactive Fixed Effects").
This code is offered with no guarantees. Not all features of this code were properly tested. Please let me know if you find any bugs or encounter any problems while using this code. All feedback is appreciated.

\strong{Linear panel regression model with weak factors:}
\itemize{
\item We consider a linear panel regression model of the form
\deqn{Y_{it} = X_{it} \beta + \sum_{k=1}^{K} Z_{k,it} \delta_{k} + \Gamma_{it} + U_{it}}
where
\itemize{
\item \eqn{Y_{it}}, \eqn{X_{i,t}}, and \eqn{Z_{k,it}} are the observed outcome variable and covariates,
\item \eqn{\Gamma_{it}} is the error component that can be correlated with \eqn{X_{it}} and \eqn{Z_{k,it}},
\item \eqn{U_{it}} is the error component modelled as a mean-zero random shock,
\item and large panel is considered, where both \eqn{N} and \eqn{T} are relatively large.
}
\item Model for \eqn{\Gamma_{it}} is referred to as a factor model with factor loadings \eqn{\lambda_{ir}} and factors \eqn{f_{tr}}, with \eqn{R} being the number of factors.
\item Other than having some strong factors, which requires \eqn{\lambda_{ir}} and \eqn{f_{tr}} to have sufficient variation across \eqn{i} and over \eqn{t}, model here allows weak factors.
\item Proposed method provides the bias-aware confidence intervals that are uniformly valid regardless of whether the factors are strong or not.
}

\strong{Debiasing approach:}
\itemize{
\item Access the preliminary estimate \eqn{\hat{\Gamma}_{pre}} along with a bound \eqn{\hat{C}} on the nuclear norm \eqn{\Vert \Gamma - \hat{\Gamma}_{pre}\Vert_{*}}.
\item Then consider the regression with the augmented outcomes \eqn{\tilde{Y}_{it} := Y_{it} - \hat{\Gamma}_{pre,it}}.
\item Construct the confidence interval with the preliminary estimate \eqn{\hat{\Gamma}_{pre}} and bound \eqn{\hat{C}} on the nuclear norm of its estimation error. Such confidence interval is bias-aware, that is, using the bound \eqn{\hat{C}} will take any remaining bias into account after the previous debias procedure.
\item For detailed implementation, please refer to Armstrong, Weidner, and Zeleneev (2024, "Robust Estimation and Inference in Panels with Interactive Fixed Effects").
}
}
\note{
We assume that all provided input parameters have values and dimensions as described above.
}
\examples{
dt <- sample_data(N = 100, T = 20, R = 1, kappa = c(0.5))
res <- honest_weak_factors(Y = dt$Y, X = dt$X, R = dt$R,
                           Gamma_LS = NULL, alpha = 0.05)
sum_res <- summary(res)
print(sum_res)

}
\references{
For a description of the model see Armstrong, Weidner, and Zeleneev (2024, "Robust Estimation and Inference in Panels with Interactive Fixed Effects").
}
